refactor: formatted components and sidebar
