+++ app/_components/Sidebar/SideBarList/ConnectClusterSidebar.tsx
@@ -6,10 +6,10 @@ export default function ConnectClusterSideBar() {
   return (
     <>
       <Link
-        href="/dashboard/connect"
+        href="/dashboard/connect-cluster"
         className={
-          pathname === "/connectCluster"
-            ? "font-semibold"
+          pathname === "/dashboard/connect-cluster"
+            ? "flex items-center p-2 text-base font-semibold text-NotikubeRed bg-gray-100 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700 group"
             : "flex items-center p-2 text-base font-normal text-gray-900 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700 group"
         }
       >
