+++ app/_components/Sidebar/SideBarList/IncidentsSidebar.tsx
@@ -1,14 +1,15 @@
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-export default function AlertsSideBar() {
+
+export default function IncidentsSideBar() {
   const pathname = usePathname();
   return (
     <>
       <Link
         href="/dashboard/incidents"
         className={
-          pathname === "/alerts"
-            ? "font-semibold"
+          pathname === "/dashboard/incidents"
+            ? "flex items-center p-2 text-base font-semibold text-NotikubeRed bg-gray-100 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700 group"
             : "flex items-center p-2 text-base font-normal text-gray-900 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700 group"
         }
       >
@@ -28,7 +29,7 @@ export default function AlertsSideBar() {
             />
           </g>
         </svg>
-        <span className="ml-3">Alerts</span>
+        <span className="ml-3">Incidents</span>
       </Link>
     </>
   );
