+++ app/_components/Sidebar/SideBarList/ProfileSidebar.tsx
@@ -8,8 +8,8 @@ export default function ProfileeSideBar() {
       <Link
         href="/dashboard/profile"
         className={
-          pathname === "/profile"
-            ? "font-semibold"
+          pathname === "/dashboard/profile"
+            ? "flex items-center p-2 text-base font-semibold text-NotikubeRed bg-gray-100 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700 group"
             : "flex items-center p-2 text-base font-normal text-gray-900 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700 group"
         }
       >
