+++ app/_components/SignUpPage.tsx
@@ -3,8 +3,8 @@
 import Link from "next/link";
 import Image from "next/image";
 import { useState } from "react";
-import SignUpImage from "../../public/NotiKubeSignUp.svg";
-import Logo from "../../public/logo.svg";
+import SignUpImage from "../../public/assets/NotiKubeSignUp.svg";
+import Logo from "../../public/assets/logo.svg";
 
 const Signup = () => {
   const [fullName, setFullName] = useState("");
