+++ app/api/get-cluster/route.ts
@@ -5,10 +5,10 @@ export default async function GET() {
 
   
   try {
-    let res = await sql`SELECT FROM clusters WHERE userid=${userid}`
+    // let res = await sql`SELECT FROM clusters WHERE userid=${userid}`
 
-    let userCluster = res
-    return NextResponse.json({userCluster})
+    // let userCluster = res
+    // return NextResponse.json({userCluster})
   } catch (err) {
     console.error('Error. Not able to get users clusters', err)
   }
