+++ package.json
@@ -22,7 +22,7 @@
     "dotenv": "^16.3.1",
     "flowbite": "^2.2.1",
     "flowbite-react": "^0.7.2",
-    "next": "^14.0.4",
+    "next": "^14.1.0",
     "next-auth": "^4.24.5",
     "node": "^21.2.0",
     "passport": "^0.7.0",
